drawing_agent/requirements.txt
# Индекс для CPU версий
--extra-index-url https://download.pytorch.org/whl/cpu

# API Server
fastapi
uvicorn
python-multipart

# LLM & Graph
langchain>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.50
openai>=1.50.0
pydantic>=2.9.0

# ML & Computer Vision
torch
torchvision
numpy>=1.26.0
Pillow>=10.4.0
scikit-learn>=1.5.0
scikit-image>=0.24.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
ultralytics>=8.2.0

# OCR & Docs
pdf2image>=1.17.0
easyocr>=1.7.0

# Infrastructure
omegaconf>=2.3.0
hydra-core>=1.3.0
python-dotenv>=1.0.0
clearml>=1.16.0
celery>=5.3.0
redis>=5.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0